# Fix error message for gettz when passed bytes