@labrys 2019-07-09
## Results of timeit
### Initial Test
```python
python -m timeit -n 1000 -s "from dateutil.tz import gettz; gettz('America/New_York')"

# Before change
1000 loops, best of 5: 3.76 usec per loop
:0: UserWarning: The test results are likely unreliable. The worst time (241 usec) was more than four times slower than the best time (3.76 usec).

# After change
1000 loops, best of 5: 3.78 usec per loop
:0: UserWarning: The test results are likely unreliable. The worst time (204 usec) was more than four times slower than the best time (3.78 usec).
```
### Optimized test
The initial test experiences a slow initial call (i believe during loading of tzfile) so I optimized it to only count subsequent gettz calls.
```python
python -m timeit -n 1000 -s "from dateutil.tz import gettz; gettz('America/Los_Angeles')" "gettz('America/New_York')"

# Before change
1000 loops, best of 5: 2.42 usec per loop

# After change
1000 loops, best of 5: 2.28 usec per loop
```

## Conclusion
The timeit differences are statistically insignificant