@pganssle 2019-11-02
@labrys Sorry for the delay in the review here, this mostly looks good, I've gone ahead and rebased this and pushed a few changes to it. We generally don't use `xfail` to enforce that a given exception is raised (instead use `pytest.raises`. Instead, we use `xfail` to write tests that *should* pass but don't because there's a bug (that way you get the regression test imposed immediately when you fix the bug).

I've gone ahead and refactored your tests. I also think that your example of `object()` being passed to `gettz` should probably also raise `TypeError`, since the `AttributeError` is exposing an arbitrary implementation detail, but I'll leave that for a different PR to fix.